Move abstractmcmc.jl to external_sampler.jl
