+++ test/runtests.jl
@@ -54,7 +54,7 @@ end
             @timeit_include("mcmc/hmc.jl")
             @timeit_include("mcmc/Inference.jl")
             @timeit_include("mcmc/sghmc.jl")
-            @timeit_include("mcmc/abstractmcmc.jl")
+            @timeit_include("mcmc/external_sampler.jl")
             @timeit_include("mcmc/mh.jl")
             @timeit_include("ext/dynamichmc.jl")
             @timeit_include("mcmc/repeat_sampler.jl")
